gamemode and handler abstraction
